delete user works on replicant (with tests!)

+++ test/func/users/delete_id.coffee
@@ -3,17 +3,17 @@ async = require('async')
 util = require('util')
 request = require('request')
 
-{nano, nanoAdmin, dbUrl} = require('config')
-{getUserDbName, hash} = require('lib/helpers')
+config = require('config')
+h = require('lib/helpers')
 
 
 describe 'DELETE /users/:id', () ->
 
   ## simple test - for now should just 403 (forbidden)
 
-  _username = hash('deleteuser@test.com')
-  _userId = 'deleteuser'
+  _username = h.hash('deleteuser@test.com')
   _password = 'deletepass'
+  _userId = 'deleteuser'
   _ctime = _mtime = 12345
   _userDoc =
     _id: _userId
@@ -22,16 +22,20 @@ describe 'DELETE /users/:id', () ->
     ctime: _ctime
     mtime: _mtime
     foo: 'delete bar'
-  cookie = null
+  _userCookie = null
   couchUser = "org.couchdb.user:#{_username}"
 
-  mainDb = nanoAdmin.db.use('lifeswap')
-  usersDb = nanoAdmin.db.use('_users')
-  userDbName = getUserDbName(userId: _userId)
+  _adminName = h.hash('tester@test.com')
+  _adminPass = 'tester'
+  _adminCookie = null
+
+  mainDb = config.nanoAdmin.db.use('lifeswap')
+  usersDb = config.nanoAdmin.db.use('_users')
+  userDbName = h.getUserDbName(userId: _userId)
 
   before (ready) ->
     ## start webserver
-    app = require('../../../app')
+    app = require('app')
     ## insert user
     insertUser = (callback) ->
       async.parallel [
@@ -52,35 +56,44 @@ describe 'DELETE /users/:id', () ->
             _userDoc._rev = res.rev
             cb()
         (cb) ->
-          nanoAdmin.db.create userDbName, (err, res) ->
+          config.nanoAdmin.db.create userDbName, (err, res) ->
             should.not.exist(err)
             cb()
       ], callback
 
     ## authenticate user
-    authUser = (callback) ->
-      nano.auth _username, _password, (err, body, headers) ->
-        should.not.exist(err)
-        should.exist(headers and headers['set-cookie'])
-        cookie = headers['set-cookie'][0]
-        callback()
+    authUsers = (callback) ->
+      async.parallel [
+        (cb) ->
+          config.nano.auth _username, _password, (err, body, hdr) ->
+            should.not.exist(err)
+            should.exist(hdr and hdr['set-cookie'])
+            _userCookie = hdr['set-cookie'][0]
+            cb()
+        (cb) ->
+          config.nano.auth _adminName, _adminPass, (err, body, hdr) ->
+            should.not.exist(err)
+            should.exist(hdr and hdr['set-cookie'])
+            _adminCookie = hdr['set-cookie'][0]
+            cb()
+      ], callback
 
-    async.series([insertUser, authUser], ready)
+    async.series([insertUser, authUsers], ready)
 
 
   after (finished) ->
     destroyUser = (callback) ->
       usersDb.get couchUser, (err, userDoc) ->
-        should.not.exist(err)
+        return callback() if err?   # should error
         usersDb.destroy(couchUser, userDoc._rev, callback)
     destroyLifeswapUser = (callback) ->
       mainDb.get _userId, (err, userDoc) ->
-        should.not.exist(err)
+        return callback() if err?   # should error
         mainDb.destroy(_userId, userDoc._rev, callback)
     destroyUserDb = (callback) ->
-      nanoAdmin.db.list (err, dbs) ->
-        dbs.should.include(userDbName)
-        nanoAdmin.db.destroy userDbName, (err, res) ->
+      config.nanoAdmin.db.list (err, dbs) ->
+        return callback() if not (userDbName in dbs)  # should callback
+        config.nanoAdmin.db.destroy userDbName, (err, res) ->
           should.not.exist(err)
           callback()
 
@@ -91,28 +104,64 @@ describe 'DELETE /users/:id', () ->
     ], finished
 
 
-  it 'should return a 403 (forbidden)', (done) ->
-    opts =
-      method: 'DELETE'
-      url: "http://localhost:3001/users/#{_userId}"
-      json: true
-      headers: {cookie}
-    request opts, (err, res, body) ->
-      should.not.exist(err)
-      res.should.have.property('statusCode', 403)
-      done()
+  describe 'regular user', () ->
+    it 'should return a 403 (forbidden)', (done) ->
+      opts =
+        method: 'DELETE'
+        url: "http://localhost:3001/users/#{_userId}"
+        json: true
+        headers: cookie: _userCookie
+      request opts, (err, res, body) ->
+        should.not.exist(err)
+        res.should.have.property('statusCode', 403)
+        done()
 
 
-  it 'should not delete _users entry', (done) ->
-    couchUser = "org.couchdb.user:#{_username}"
-    usersDb.get couchUser, (err, userDoc) ->
-      should.not.exist(err)
-      userDoc.should.have.property('_id', couchUser)
-      done()
+    it 'should not delete _users entry', (done) ->
+      usersDb.get couchUser, (err, userDoc) ->
+        should.not.exist(err)
+        userDoc.should.have.property('_id', couchUser)
+        done()
 
 
-  it 'should not delete \'user\' type entry in lifeswap db', (done) ->
-    mainDb.get _userId, (err, userDoc) ->
-      should.not.exist(err)
-      userDoc.should.eql(_userDoc)
-      done()
+    it 'should not delete \'user\' type entry in lifeswap db', (done) ->
+      mainDb.get _userId, (err, userDoc) ->
+        should.not.exist(err)
+        userDoc.should.eql(_userDoc)
+        done()
+
+    it 'should not delete user DB', (done) ->
+      config.nanoAdmin.db.list (err, dbs) ->
+        should.not.exist(err)
+        dbs.should.include(userDbName)
+        done()
+
+  describe 'constable', () ->
+
+    it 'should return a 200 (OK)', (done) ->
+      opts =
+        method: 'DELETE'
+        url: "http://localhost:3001/users/#{_userId}"
+        json: true
+        headers: cookie: _adminCookie
+      request opts, (err, res, body) ->
+        should.not.exist(err)
+        res.should.have.property('statusCode', 200)
+        done()
+
+    it 'should delete the _users entry', (done) ->
+      usersDb.get couchUser, (err, userDoc) ->
+        should.exist(err)
+        err.should.have.property('status_code', 404)
+        done()
+
+    it 'should delete the \'user\' type entry in lifeswap db', (done) ->
+      mainDb.get _userId, (err, userDoc) ->
+        should.exist(err)
+        err.should.have.property('status_code', 404)
+        done()
+
+    it 'should delete the user DB', (done) ->
+      config.nanoAdmin.db.list (err, dbs) ->
+        dbs.should.not.include(userDbName)
+        done()
